Isolate filter tests
